fix redux logic
