+++ src/components/CustomInput/CustomInput.js
@@ -3,36 +3,19 @@ import classNames from "classnames";
 import styles from "./CustomInput.module.scss";
 import { api } from "../../axios";
 import { useSelector, useDispatch } from "react-redux";
-import { complete } from "../../store/resultSlice";
+import { searchResultRequest } from "../../redux/resultSlice";
 
 const { inputBox, inputContent, inputText, buttonBox } = styles;
 
-const ONE_MINUTE = 1000 * 60;
-
 const CustomInput = ({ placeholder, value, disable }) => {
-  const resultLists = useSelector((state) => state.result.value);
   const dispatch = useDispatch();
+  const resultLists = useSelector((state) => state.resultSlice);
 
   const handleInputChange = (e) => {
     const searchInput = e.target.value;
 
     if (searchInput) {
-      const checkCache = localStorage.getItem(searchInput);
-
-      if (checkCache) {
-        dispatch(complete(checkCache));
-      } else {
-        api.get(searchInput).then((response) => {
-          // console.log(response.data);
-
-          const object = {
-            data: response.data,
-            expireTime: new Date().getTime() + ONE_MINUTE,
-          };
-
-          localStorage.setItem(searchInput, JSON.stringify(object));
-        });
-      }
+      dispatch(searchResultRequest(searchInput));
     }
   };
