--- src/store/store.js
@@ -1,8 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-import resultReducer from "./resultSlice";
-
-export default configureStore({
-  reducer: {
-    result: resultReducer,
-  },
-});
